Changes for the headline structure
The readme had an incoherent headline structure which made the comprehension a bit difficult.  It was cumbersome to distinguish between the `vagrant` part and the `docker` part due to the fact the `vagrant` was an `h1` while `docker` was an `h2`. and right before the docker part there were the `configuration` and `usage` parts which were also an `h1` making the impression that all steps might be necessary even though you just want to use the `docker` part. instead they are two complete different processes. therefor i would suggest make the vagrant and docker headlines each a `h2` so the only `h1` in the document is the `visual regression testing made easy`. and all the sub headlines of `vagrant` and `docker`  a `h3`. that would be a quick and easy fix to make the readme a bit easier to comprehend.